requirements.txt
pycryptodomex>=3.4.7
# see https://github.com/Tanganelli/CoAPthon3/issues/29
CoAPthon3 @ git+https://github.com/Tanganelli/CoAPthon3@89d5173
pyyaml>=5.4.1
flask>=2.0.0
waitress>=2.0.0
Cerberus>=1.3.4